refactor(data-model): move primitive codecs into `codec-common`
The standalone codecs for JS primitives (`BigIntCodec`,
`SpecialNumberCodec`, `SymbolCodec`, `UndefinedCodec`) -- which have no
owned class to host a `[CODEC]` -- move from `codec-json` into
`codec-common` alongside `BaseFabricCodec` and the codec vocabulary, with
their tests. They're now exported from the `codec-common` barrel; their
codec-common sibling imports become relative, and `createDefaultRegistry`
(still in `codec-json`) pulls them from `codec-common`.
