+++ packages/data-model/src/codec-common/SpecialNumberCodec.ts
