+++ packages/data-model/src/codec-common/index.ts
@@ -15,3 +15,9 @@ export {
   EMPTY_RECONSTRUCTION_CONTEXT,
   EmptyReconstructionContext,
 } from "./EmptyReconstructionContext.ts";
+
+// Standalone codecs for JS primitives (no owned class to host a `[CODEC]`).
+export { UndefinedCodec } from "./UndefinedCodec.ts";
+export { BigIntCodec } from "./BigIntCodec.ts";
+export { SpecialNumberCodec } from "./SpecialNumberCodec.ts";
+export { SymbolCodec } from "./SymbolCodec.ts";
