+++ packages/data-model/src/codec-json/index.ts
@@ -19,9 +19,3 @@ export type { JsonWireValue } from "./interface.ts";
 // Codec registry and factory.
 export { CodecRegistry } from "./CodecRegistry.ts";
 export { createDefaultRegistry } from "./createDefaultRegistry.ts";
-
-// Standalone codecs for JS primitives (no owned class to host a `[CODEC]`).
-export { UndefinedCodec } from "./UndefinedCodec.ts";
-export { BigIntCodec } from "./BigIntCodec.ts";
-export { SpecialNumberCodec } from "./SpecialNumberCodec.ts";
-export { SymbolCodec } from "./SymbolCodec.ts";
