+++ packages/data-model/test/codec-common/UndefinedCodec.test.ts
@@ -1,7 +1,7 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 
-import { UndefinedCodec } from "@/codec-json/UndefinedCodec.ts";
+import { UndefinedCodec } from "@/codec-common/UndefinedCodec.ts";
 import { CODEC_TYPE_TAGS } from "@/codec-common/codec-type-tags.ts";
 import { EMPTY_RECONSTRUCTION_CONTEXT } from "@/codec-common/EmptyReconstructionContext.ts";
